Changed the api to check if the book is checked out by user; used request param instead of request body.

+++ backend/OnlineLibrary/src/main/java/com/arka99/OnlineLibrary/dao/CheckoutRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Repository
 public interface CheckoutRepository extends JpaRepository<Checkout, Long> {
 
-    Boolean existsCheckoutByUserEmailAndBookId(String userEmail, Long bookId);
+    Boolean existsCheckoutByUserEmailAndBookId(String userEmail, Integer bookId);
 
     List<Checkout> findCheckoutsByUserEmail(String userEmail);
 }
